requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
pydantic==1.10.13
typing-extensions==4.9.0
aiohttp==3.9.1
ruff==0.1.9
openai==0.28.1
